test(umbp): consolidate redundant distributed test coverage
Merge locality routing coverage into the main RouteGet suite and remove cases that only retest the mock backend or already-covered selection paths. Rename the remaining registry suite for its actual subject and shorten the metrics reporting interval so buffering tests do not wait on the production cadence.
