+++ tests/cpp/umbp/distributed/test_backend_registry.cpp
@@ -20,11 +20,9 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-// Proves BackendRegistry dispatch across two distinct MediumBackend
-// implementations (design doc §5 Phase 2, default option (b)): registering a
-// MockBackend alongside a real one and driving both purely through the
-// registry / MediumBackend interface, with no concrete type named at the call
-// site.
+// BackendRegistry dispatch and cross-backend event aggregation. MockBackend is
+// only the test double; its own behavior is exercised transitively here and by
+// the PeerPool / PeerService suites.
 
 #include <gtest/gtest.h>
 
@@ -36,51 +34,6 @@
 #include "umbp/distributed/peer/backend/mock_backend.h"
 
 namespace mori::umbp {
-
-TEST(MockBackend, ZeroCapacityByDesign) {
-  MockBackend b(TierType::HBM);
-  EXPECT_EQ(b.Capacity().total_bytes, 0u);
-  EXPECT_EQ(b.Capacity().available_bytes, 0u);
-}
-
-TEST(MockBackend, AllocateCommitResolveEvictRoundTrip) {
-  MockBackend b(TierType::HBM);
-  ASSERT_TRUE(b.Init(nullptr));
-
-  auto allocated = b.BatchAllocate({{"k", 128}});
-  ASSERT_EQ(allocated.size(), 1u);
-  ASSERT_EQ(allocated[0].outcome, AllocateOutcome::kSuccessAllocated);
-
-  auto committed = b.BatchCommit({{allocated[0].slot_id, "k"}});
-  ASSERT_EQ(committed.size(), 1u);
-  EXPECT_TRUE(committed[0].success);
-  EXPECT_EQ(committed[0].bytes_committed, 128u);
-  EXPECT_EQ(b.OwnedKeyCount(), 1u);
-
-  auto resolved = b.BatchResolve({"k", "missing"}, /*include_descs=*/true);
-  ASSERT_EQ(resolved.size(), 2u);
-  EXPECT_TRUE(resolved[0].found);
-  EXPECT_EQ(resolved[0].size, 128u);
-  EXPECT_FALSE(resolved[1].found);
-
-  auto evicted = b.Evict({"k"});
-  ASSERT_EQ(evicted.size(), 1u);
-  EXPECT_EQ(evicted[0].bytes_freed, 128u);
-  EXPECT_EQ(b.OwnedKeyCount(), 0u);
-}
-
-TEST(MockBackend, DrainAndSnapshotEvents) {
-  MockBackend b(TierType::HBM);
-  auto allocated = b.BatchAllocate({{"a", 8}, {"b", 8}});
-  b.BatchCommit({{allocated[0].slot_id, "a"}, {allocated[1].slot_id, "b"}});
-
-  auto drained = b.DrainPendingEvents();
-  EXPECT_EQ(drained.size(), 2u);
-  EXPECT_TRUE(b.DrainPendingEvents().empty());  // outbox cleared
-
-  auto snap = b.SnapshotOwnedKeys();
-  EXPECT_EQ(snap.size(), 2u);
-}
 
 // The registry dispatches to whichever concrete backend is registered for a
 // tier, without the caller naming MockBackend anywhere below this line — this
@@ -106,10 +59,8 @@ TEST(BackendRegistryDispatch, RoutesByTierThroughTheInterfaceOnly) {
 
 TEST(BackendRegistryDispatch, NamedSameTierInstancesKeepDenseStableIds) {
   BackendRegistry registry;
-  ASSERT_TRUE(
-      registry.Register("dram-primary", std::make_unique<MockBackend>(TierType::DRAM)));
-  ASSERT_TRUE(
-      registry.Register("dram-secondary", std::make_unique<MockBackend>(TierType::DRAM)));
+  ASSERT_TRUE(registry.Register("dram-primary", std::make_unique<MockBackend>(TierType::DRAM)));
+  ASSERT_TRUE(registry.Register("dram-secondary", std::make_unique<MockBackend>(TierType::DRAM)));
 
   ASSERT_EQ(registry.Size(), 2u);
   EXPECT_EQ(registry.Get(TierType::DRAM), registry.Get("dram-primary"));
@@ -118,11 +69,9 @@ TEST(BackendRegistryDispatch, NamedSameTierInstancesKeepDenseStableIds) {
   EXPECT_EQ(registry.BackendId(registry.Get("dram-secondary")), 1u);
 
   auto allocated = registry.Get("dram-primary")->BatchAllocate({{"old-state", 8}});
-  registry.Get("dram-primary")
-      ->BatchCommit({{allocated.front().slot_id, "old-state"}});
+  registry.Get("dram-primary")->BatchCommit({{allocated.front().slot_id, "old-state"}});
   ASSERT_EQ(registry.Get("dram-primary")->OwnedKeyCount(), 1u);
-  ASSERT_TRUE(
-      registry.Register("dram-primary", std::make_unique<MockBackend>(TierType::DRAM)));
+  ASSERT_TRUE(registry.Register("dram-primary", std::make_unique<MockBackend>(TierType::DRAM)));
   EXPECT_EQ(registry.Get("dram-primary")->OwnedKeyCount(), 0u);
   EXPECT_EQ(registry.Size(), 2u);
   EXPECT_EQ(registry.BackendId(registry.Get("dram-primary")), 0u);
